Finally it all should work, I tested it outside the program but not yet with anything we want to monitor. it saves all stats in *.png files at the moment

+++ src/abalone/exec/StatisticGenerator.java
@@ -5,7 +5,7 @@ public interface StatisticGenerator
 	/**
 	 * A method to convert the current state of the object to some long
 	 */
-	long getCurrentState();
+	double getCurrentState();
 	
 	//TODO: make sure that one class can save multiple data about itself
 }
